Create message.txt
Write a JavaScript script that validates all user entries for which it makes sense to do so in the form that permits your customers to choose products and/or services. Be sure to inform the user if any data entry is invalid with an appropriate popup message. Once all input has been validated, display to your customer some relevant output related to the data that has been entered. A typical display should include a summary of the items ordered, the quantity of each item, the cost of each item, and the total cost.



Note:

1, You can use our previous exercises on this.

2, Send in email your screenshot of the outputs: forms, css and js scripts.

3. Make your form simple and with css styles.